requirements.txt
# Web scraping dependencies
requests==2.31.0
beautifulsoup4==4.12.3

# Google Calendar API dependencies
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Additional parsing utilities
lxml==5.1.0
